Restore main activity
