+++ app/src/main/res/layout/fragment_recently_read.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
@@ -13,4 +13,4 @@
 
     </android.support.v7.widget.RecyclerView>
 
-</RelativeLayout>
+</FrameLayout>
